git-svn-id: http://deepneko.dyndns.org/svn/repos/explorer@399 76beb010-293a-44e9-8231-3d3008485057

+++ bin/updatedb.rb
@@ -28,7 +28,7 @@ alldate = $con.execute("select date from filelist").flatten
 
 # if file doesn't exist, delete from database
 allpath.each do |path,flv|
-  if !File.exists?(path)
+  if path && !File.exists?(path)
     print "delete:" + path + "\n"
     begin
       $con.execute("delete from filelist where path=\"#{path}\"")
